FIX [ objload ] face reading only worked for v/t/n
in case of "v" or "v/t" (which are valid obj face formats) the reading of last vertex in the face produced an error (cause tried to peek at eof) which resulted in the last vertex being discarded
therefore every face missed one vertex. for triangular faces that meant the whole face was invalid an discard. so triangular meshes got discarded completly.